requirements.txt
Django==5.1.2
django-filter==24.3